adapting how upload and download work
