+++ test/failures_testsuite/base_test.py
@@ -83,7 +83,6 @@ class BaseTest(TestCase):
         :return:
         """
         self = cls()
-        self._delete_directory(directory='tmp')
 
     def setUp(self):
         self.s3 = self.s3_controller.s3[self.s3_service_name]
@@ -93,127 +92,3 @@ class BaseTest(TestCase):
 
     def tearDown(self):
         pass
-
-    def upload_file(self):
-        """
-         - Create random 2M file
-         - Calc its md5 checksum hash
-         - Rename file to make its name = md5
-         - Upload it
-        :return: file_name
-        """
-        self.minio['name'] = 'minio_{}'.format(self.s3_service_name)
-        self.minio['bucket'] = 'bucket{}'.format(self.s3_service_name)
-        self._create_directory(directory='tmp')
-        self.file_name = self._create_file(directory='tmp', size=1024 * 1024 * 2)
-
-        self._check_minio_ip()
-
-        config_minio_cmd = '/bin/mc config host add {} {} {} {}'.format(self.minio['name'],
-                                                                        self.minio['minio_ip'],
-                                                                        self.minio['username'],
-                                                                        self.minio['password'])
-        self.logger.info('config minio')
-        self.logger.info(config_minio_cmd)
-        out, err = self.execute_cmd(cmd=config_minio_cmd)
-        if err:
-            self.logger.error(err)
-
-        self.logger.info('create {} bucket'.format(self.minio['bucket']))
-        create_bucket_cmd = '/bin/mc mb {}/{}'.format(self.minio['name'], self.minio['bucket'])
-        out, err = self.execute_cmd(cmd=create_bucket_cmd)
-        if err:
-            self.logger.warning(err)
-
-        self.logger.info('Upload {} to  {} bucket'.format(self.file_name, self.minio['bucket']))
-        err = self._upload_file(self.minio['name'], self.minio['bucket'], 'tmp/{}'.format(self.file_name))
-        if err:
-            return None
-
-        self.logger.info('{} file has been uploaded'.format(self.file_name))
-        return self.file_name
-
-    def _upload_file(self, minio, bucket, file_name):
-        upload_cmd = '/bin/mc cp {} {}/{}/{}'.format(file_name, minio, bucket, file_name)
-        out, err = self.execute_cmd(cmd=upload_cmd)
-        if err:
-            self.logger.error(err)
-            return err
-
-        return
-
-    def download_file(self, file_name, keep_trying=False):
-        """
-         - downlaod file
-         - return its md5 checksum hash
-        :return: str(downloaded_file_md5)
-        """
-        self.logger.info('Download {} .... '.format(file_name))
-        download_cmd = '/bin/mc cp {}/{}/tmp/{} tmp/{}_out'.format(self.minio['name'], self.minio['bucket'],
-                                                                   file_name, file_name)
-        out, err = self.execute_cmd(cmd=download_cmd)
-        if keep_trying and err:
-            for _ in range(50):
-                out, err = self.execute_cmd(cmd=download_cmd)
-                if not err:
-                    break
-                else:
-                    time.sleep(5)
-
-        if err:
-            self.logger.error(err)
-            return None
-
-        return self.calc_md5_checksum('tmp/{}_out'.format(file_name))
-
-    def get_s3_info(self):
-        self.s3_data = self.s3.service.data['data']
-        self.parity = self.s3_data['parityShards']
-        self.shards = self.s3_data['dataShards']
-        self.logger.info('- shards {}, parity {}'.format(self.shards, self.parity))
-        self.minio = {'minio_ip': self.s3_data['minioUrls']['public'],
-                      'username': self.s3_data['minioLogin'],
-                      'password': self.s3_data['minioPassword']}
-        self.tlog = self.s3_data['tlog']
-
-    def execute_cmd(self, cmd):
-        sub = Popen(cmd, stdout=PIPE, stderr=PIPE, shell=True)
-        out, err = sub.communicate()
-        return out, err
-
-    def calc_md5_checksum(self, file_path):
-        hash_md5 = hashlib.md5()
-        with open(file_path, "rb") as f:
-            for chunk in iter(lambda: f.read(4096), b""):
-                hash_md5.update(chunk)
-        return hash_md5.hexdigest()
-
-    def _create_directory(self, directory):
-        if not os.path.exists(directory):
-            os.makedirs(directory)
-
-    def _delete_directory(self, directory):
-        os.system("rm -rf {}".format(directory))
-
-    def _create_file(self, directory, size):
-        with open('{}/random'.format(directory), 'wb') as fout:
-            fout.write(os.urandom(size))  # 1
-
-        file_name = self.calc_md5_checksum('{}/random'.format(directory))
-
-        os.rename('{}/random'.format(directory), '{}/{}'.format(directory, file_name))
-        return file_name
-
-    def _delete_file(self, file_path):
-        os.system("rm -f {}".format(file_path))
-
-    def _check_minio_ip(self):
-        self.logger.info('Checking if minio ip is existing')
-        for _ in range(5):
-            if 'http' in self.minio['minio_ip']:
-                break
-            else:
-                time.sleep(60)
-                self.logger.info('Wait ... ')
-        else:
-            raise RuntimeError("There is no minio ip, We can't connect to it!")
