+++ test/failures_testsuite/testcases/test01_tlog.py
@@ -22,8 +22,7 @@ class TestTlog(BaseTest):
          - Start tlog
          - Download file, should succeed
         """
-        self.file_name = self.upload_file()
-        md5_before = self.file_name
+        file_name, bucket_name, md5_before = self.s3.upload_file()
 
         self.s3.failures.tlog_down()
         time.sleep(60)
@@ -41,7 +40,7 @@ class TestTlog(BaseTest):
         else:
             self.assertTrue(self.s3.failures.tlog_status())
 
-        md5_after = self.download_file(file_name=self.file_name, keep_trying=True)
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
 
     @skip('https://github.com/threefoldtech/0-templates/issues/179')
@@ -54,7 +53,9 @@ class TestTlog(BaseTest):
         """
         self.s3.failures.tlog_down()
         time.sleep(60)
-        self.assertIsNone(self.upload_file())
+        with self.assertRaises(RuntimeError):
+            self.s3.upload_file()
+            self.logger.info("Can't upload the file")
 
         self.s3.failures.tlog_up()
         for _ in range(10):
@@ -66,10 +67,9 @@ class TestTlog(BaseTest):
         else:
             self.assertTrue(self.s3.failures.tlog_status())
 
-        self.file_name = self.upload_file()
-        md5_before = self.file_name
+        file_name, bucket_name, md5_before = self.s3.upload_file()
 
-        md5_after = self.download_file(file_name=self.file_name, keep_trying=True)
+        md5_after = self.s3.download_file(file_name, bucket_name, timeout=250)
         self.assertEqual(md5_after, md5_before)
 
     def test003_upload_kill_tlog_download(self):
@@ -79,13 +79,12 @@ class TestTlog(BaseTest):
          - wait 60 sec, tlog should be returned
          - Download file, should succeed
         """
-        self.file_name = self.upload_file()
-        md5_before = self.file_name
+        file_name, bucket_name, md5_before = self.s3.upload_file()
 
         self.assertFalse(self.s3.failures.kill_tlog())
         time.sleep(60)
 
-        md5_after = self.download_file(file_name=self.file_name, keep_trying=True)
+        md5_after = self.s3.download_file(file_name, bucket_name, timeout=250)
         self.assertEqual(md5_after, md5_before)
 
     def test004_kill_tlog_upload_download(self):
@@ -98,8 +97,7 @@ class TestTlog(BaseTest):
         self.assertFalse(self.s3.failures.kill_tlog())
         time.sleep(60)
 
-        self.file_name = self.upload_file()
-        md5_before = self.file_name
+        file_name, bucket_name, md5_before = self.s3.upload_file()
 
-        md5_after = self.download_file(file_name=self.file_name, keep_trying=True)
+        md5_after = self.s3.download_file(file_name, bucket_name, timeout=250)
         self.assertEqual(md5_after, md5_before)
