+++ test/failures_testsuite/testcases/test02_zdb.py
@@ -2,7 +2,7 @@ from random import randint
 from base_test import BaseTest
 import unittest, time
 
-class ZDBFailures(BaseTest):    
+class ZDBFailures(BaseTest):
     def tearDown(self):
         super().tearDown()
 
@@ -12,14 +12,14 @@ class ZDBFailures(BaseTest):
         - kill zdb process  and make sure it will restart automatically.
         - download uploaded file, should pass
         """
-        md5_before = self.upload_file()
+        file_name, bucket_name, md5_before = self.s3.upload_file()
 
         self.logger.info('kill zdb process and make sure it will restart automatically')
         flag = self.s3.failures.zdb_process_down()
         self.assertTrue(flag, "zdb didn't restart")
 
         self.logger.info("Download uploaded file, and check that both are same.")
-        md5_after = self.download_file(file_name=md5_before)
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
 
     def test002_upload_stop_parity_zdb_download(self):
@@ -33,17 +33,14 @@ class ZDBFailures(BaseTest):
         - assert md5 checksum is matching
         - Start n zdb
         """
-        self.file_name = self.upload_file()
-        md5_before = self.file_name
-
-        md5_after = self.download_file(file_name=self.file_name)
+        file_name, bucket_name, md5_before = self.s3.upload_file()
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
-        self._delete_file('tmp/{}'.format(md5_after))
 
         self.logger.info('Stop {} zdb'.format((self.parity)))
         self.s3.failures.zdb_down(count=self.parity)
 
-        md5_after = self.download_file(file_name=self.file_name)
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
 
         self.logger.info('Start {} zdb'.format((self.parity)))
@@ -57,12 +54,11 @@ class ZDBFailures(BaseTest):
         - assert md5 checksum is matching
         - start n zdb, should pass
         """
-        self.file_name = self.upload_file()
+        file_name, bucket_name, md5_before = self.s3.upload_file()
         self.logger.info(' Stop {} zdb'.format((self.parity)))
-        md5_before = self.file_name
         self.s3.failures.zdb_down(count=self.parity)
 
-        md5_after = self.download_file(file_name=self.file_name)
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
 
         self.logger.info(' Start {} zdb'.format((self.parity)))
@@ -76,15 +72,14 @@ class ZDBFailures(BaseTest):
         - download file, should pass
         - assert md5 checksum is matching
         """
-        self.file_name = self.upload_file()
+        file_name, bucket_name, md5_before = self.s3.upload_file()
         self.logger.info(' Stop {} zdb'.format((self.parity)))
-        md5_before = self.file_name
         self.s3.failures.zdb_down(count=self.parity)
 
         self.logger.info(' Start {} zdb'.format((self.parity)))
         self.s3.failures.zdb_up(count=self.parity)
 
-        md5_after = self.download_file(file_name=self.file_name)
+        md5_after = self.s3.download_file(file_name, bucket_name)
         self.assertEqual(md5_after, md5_before)
 
     def test005_stop_greater_parity_zdb_upload(self):
@@ -95,16 +90,14 @@ class ZDBFailures(BaseTest):
         - Download the uploaded file, should succeed
         - Start n+ zdb
         """
-        self.file_name = self.upload_file()
+        self.s3.upload_file()
         zdb_turn_down = self.parity + randint(1, self.shards)
         self.logger.info(' Stop {} zdb'.format(zdb_turn_down))
         self.s3.failures.zdb_down(count=zdb_turn_down)
 
-        try:
+        self.logger.info('Uploading should raise an error')
+        with self.assertRaises(RuntimeError):
             self.upload_file()
-            self.assertTrue(False, 'Uploading should raise an error')
-        except:
-            pass
 
         self.logger.info(' Start {} zdb'.format(zdb_turn_down))
         self.s3.failures.zdb_up(count=zdb_turn_down)
