+++ test/john.txt
@@ -1,0 +1,1 @@
+john
